sync: add a progress bar and slightly improve the reliability of downloading
Previously if a download failed we'd leave a partially downloaded file
behind, which frequently crashed us when we parsed it later. Change it
so we first download into a temp file and then copy to the location it's
supposed to be.

This might help with the implementation of delta downloads later, idk